Deduplicate helpers & fix lint issues from #1099
